requirements.txt
streamlit>=1.50.0, <2.0.0
google-generativeai
openai>=2.0.0,<3.0.0
python-dotenv>=1.1.1,<2.0.0
pandas==1.5.3
pyodbc==4.0.35